Add CRUD Vendor Product

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -2,24 +2,27 @@
 
 namespace App\Http\Controllers\Backend;
 
-use App\DataTables\ProductImageGalleryDataTable;
+use App\DataTables\VendorProductImageGalleryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductImageGallery;
-use Illuminate\Http\Request;
 use App\Traits\ImageUploadTrait;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
-
-class ProductImageGalleryController extends Controller
+class VendorProductImageGalleryController extends Controller
 {
     use ImageUploadTrait;
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request, ProductImageGalleryDataTable $dataTable)
+    public function index(Request $request, VendorProductImageGalleryDataTable $dataTable)
     {
         $product = Product::findOrFail($request->product);
-        return $dataTable->render('admin.product.image-gallery.index', compact('product'));
+        if ($product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
+        return $dataTable->render('vendor.product.image-gallery.index', compact('product'));
     }
 
     /**
@@ -83,6 +86,10 @@ class ProductImageGalleryController extends Controller
     public function destroy(string $id)
     {
         $productImage = ProductImageGallery::findOrFail($id);
+        /**Check product vendor */
+        if ($productImage->product->vendor_id !== Auth::user()->vendor->id) {
+            abort(404);
+        }
         $this->deleteImage($productImage->image);
         $productImage->delete();
